requirements.txt.txt
1.matplotlib
2.strealit
3.pandas
4.numpy
5.statsmodels